v1.3, an old commit with a large file fucked everything up

+++ util.py
@@ -4,7 +4,7 @@ from random import randint
 from json import *
 from requests import get
 from locale import getlocale
-from os import mkdir, path, remove, removedirs, rename
+from os import mkdir, path, remove
 from xml.etree import ElementTree
 from xml.sax.saxutils import escape
 from multiprocessing import Process
@@ -14,6 +14,8 @@ import requests
 from subprocess import Popen
 from functools import partial
 from win10toast_click import ToastNotifier
+from win32com.client import Dispatch
+
 
 def notify_desktop(title,text):
     # initialize 
@@ -23,7 +25,7 @@ def notify_desktop(title,text):
     toaster.show_toast(
         title, # title
         text, # message 
-        icon_path=None, # 'icon_path' 
+        icon_path="static/favicon.ico", # 'icon_path' 
         duration=5, # for how many seconds toast should be visible; None = leave notification in Notification Center
         threaded=True, # True = run other code in parallel; False = code execution will wait till notification disappears 
         callback_on_click=partial(open_tab,"http://127.0.0.1:21987") # click notification to run function 
@@ -59,15 +61,24 @@ def check_updates():
 
 def emergency_redownload():
     
+    
+    notify_desktop("CopyPasta update","CopyPasta is downloading its files, it can take some time...")
+    
     if not path.exists("templates/"):
         mkdir("templates")
 
     if not path.exists("static/"):
         mkdir("static")
-        open("static/favicon.ico","w")
+        
+        f = open("static/favicon.ico","w")
+        f.close()
+
+        f = open("static/qr.jpeg","w")
+        f.close()
 
         with open("static/update.Blue","w") as f:
             f.write("1")
+            f.close()
 
         mkdir("static/dist")
         mkdir("static/dist/css")
@@ -123,9 +134,9 @@ def download_templates():
 
 
 
-def update_main_executable():
+def update_main_executable(version):
 
-    if not literal_eval(get("https://api.github.com/repos/CopyPastaOfficial/CopyPasta/tags").text)[0]['name'] == "1.2":
+    if not literal_eval(get("https://api.github.com/repos/CopyPastaOfficial/CopyPasta/tags").text)[0]['name'] == version:
         
 
         with open("copypasta(1).exe","wb") as f:
@@ -155,9 +166,11 @@ def init_history_file():
     initialize the history xml file
 
     """
-    with open("static/history.xml","w") as f:
-        f.write("<history>\n</history>")
-        f.close()
+    
+    if not path.exists("static/history.xml"):
+        with open("static/history.xml","w") as f:
+            f.write("<history>\n</history>")
+            f.close()
 
 
 def get_history():
@@ -219,3 +232,17 @@ def open_browser_if_settings_okay(url):
     
     if path.exists("static/tab"):
         Process(target=open_link_process,args=(url,)).start()
+        
+        
+        
+def create_shortcut(path, target='', wDir='', icon=''):    
+ 
+    shell = Dispatch('WScript.Shell')
+    shortcut = shell.CreateShortCut(path)
+    shortcut.Targetpath = target
+    shortcut.WorkingDirectory = wDir
+    if icon == '':
+        pass
+    else:
+        shortcut.IconLocation = icon
+    shortcut.save()
